Added files via upload
Company wants to transmit data over the telephone, but they are concerned 
that their phones may be trapped. All their data are transmitted as four digit integer.
Write a C# application that will encrypt their data so that it may be transmitted more securely. Application should read a four digit number entered by the user through input box and encrypt it as follows:
Replace each digit by ( the sum of that digit plus 7)modulus 10. Then swap the first digit with the third and swap the second digit wth the fourth. Print the encrypted integer. Write a separate application that inputs an encrypted four digit integer and decrypts it to form the original number