Added files via upload
Create a C# application that will be consisting of list box and combo box controls Fill up the combo box with names of provinces of Pakistan and when the user selects an item from the combo box,it should be removed from the combo box and add it to the list box control and names  and pictures of measure cities of this province should be displayed in the text boxes and picture box respectively.Program should check to ensure that the combo box contains at least one item. If it does not ,display a message box and terminate the program.